requirements.txt
loguru~=0.7.2
openai~=1.34.0
llama_cpp_python~=0.2.78
python-dotenv~=1.0.1
Jinja2~=3.1.4
termcolor~=2.4.0
